PPW13 - Authorization (Gate, Middleware, View) (#2)
* Addresses article text and Post factory.

* Reduced overall post from seeding

Also new category.

* Factory random

* php artisan make:migration table_users

* add is_admin role to user table

* Add route name

* Fix category link

* fix issue editing formatted text

* tidy up userpostcontroller and post thumbnail

* prevent XSS in post body

* make sure slug is valid (for link)

* Add edit and delete to post

* 'login' route

* fix variable naming error

* php artisan make:controller AdminPageController

* add route to user in post edit.blade.php

* AdminPage

* AdminPage for posts

* AdminPage add several table

* php artisan make:middleware AdminOnly

* AdminOnly middleware

* php artisan make:middleware UserOwnPost

* User can only edit or delete their own post

* Fix user can't comment on other post

* Fix user can't create new post

* Gate (Admin)

* Gate (UserOwnPost)

* Authorization on view (Admin)

* Authorization on view (UserOwnPost)

* Authorization on controller (UserOwnPost & AdminOnly)

* Better edit button color

* Error flash message

* Add message to unauthorized action.

* Delete users, posts, comments for admin

+++ app/Http/Controllers/UserPostController.php
@@ -55,8 +55,11 @@ class UserPostController extends Controller
         return redirect('/');
     }
 
-    public function edit(Post $post): View
+    public function edit(Post $post): View|RedirectResponse
     {
+        if (auth()->user()->cannot('userownpost', $post))
+            return redirect()->back()->with('error', 'You are not authorized to edit this post.');
+
         $categories = Category::all();
         return view(
             'user.posts.edit',
@@ -66,6 +69,8 @@ class UserPostController extends Controller
 
     public function update(Post $post): RedirectResponse
     {
+        if (auth()->user()->cannot('userownpost', $post))
+            return redirect()->back()->with('error', 'You are not authorized to edit this post.');
         $attributes = $this->validatePost($post);
 
         if ($attributes['thumbnail'] ?? false) {
@@ -79,6 +84,8 @@ class UserPostController extends Controller
 
     public function destroy(Post $post): RedirectResponse
     {
+        if (auth()->user()->cannot('userownpost', $post) && auth()->user()->cannot('admin'))
+            return redirect()->back()->with('error', 'You are not authorized to delete this post.');
         $post->delete();
         return redirect('/')->with('success', 'Successfully deleted post!');
     }
